+++ packages/pwa-kit-react-sdk/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## v1.4.0-dev (Jan 06, 2022)
--   Upgrade to React 17 [#278](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/278)
+-   Add React 17 support [#278](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/278)
 
 ## v1.3.0 (Jan 06, 2022)