+++ packages/pwa-kit-react-sdk/CHANGELOG.md
@@ -1,8 +1,12 @@
-## v1.3.0-dev (Nov 18, 2021)
+## v1.4.0-dev (Jan 06, 2022)
+-   Add React 17 support [#278](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/278)
+
+## v1.3.0 (Jan 06, 2022)
 
 -   Add `__server_only` and `__pretty_print` server rendering flags aliases. [#250](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/250)
 -   Do not show stack trace in remote environment windowGlobals [#230](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/230/files)
 -   `createApp` takes a new option `enableLegacyRemoteProxying` which defaults to `true`. When set to `false`, local development proxying is disabled when running remotely. In future, local development proxying will *always* be disabled when running remotely. [#205](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/205)
+-   Add `PwaKitConfigPlugin` webpack plugin. [#255](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/255)
 
 ## v1.2.0 (Nov 18, 2021)